+++ packages/react-theme-default/src/components/SCCommentsFeedObject.tsx
@@ -1,6 +1,14 @@
 const Component = {
   styleOverrides: {
     root: ({theme}: any) => ({
+      width: '100%',
+      '& .SCCommentsFeedObject-no-comments': {
+        padding: theme.spacing(),
+        paddingLeft: 0
+      },
+      '& .SCCommentsFeedObject-comment-not-found': {
+        padding: theme.spacing()
+      },
       '& .SCCommentObject-root': {
         '& > div': {
           background: 'transparent'
@@ -23,6 +31,7 @@ const Component = {
         marginBottom: theme.spacing(0.5)
       },
       '& .SCCommentObjectReply-root': {
+        background: 'transparent',
         '& .SCCommentObjectReply-comment': {
           marginBottom: theme.spacing(4)
         }