+++ packages/react-theme-default/src/components/SCNavigationToolbarMobileSkeleton.tsx
@@ -1,0 +1,12 @@
+const Component = {
+  styleOverrides: {
+    root: ({theme}: any) => ({
+      '& .SCNavigationToolbarMobileSkeleton-logo': {
+        width: 100,
+        height: 20
+      }
+    })
+  }
+};
+
+export default Component;